# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaksi Pekerjaan API",
        "description": "API untuk aplikasi **Transaksi Pekerjaan** — pencatatan arus kas per kegiatan\nbeserta perhitungan bagi hasil.\n\n### Alur autentikasi\n1. `POST /api/auth/login` dengan email + password → dapat `access_token`.\n2. Kirim `Authorization: Bearer <access_token>` untuk semua endpoint lain.\n3. Opsional: `POST /api/auth/biometric/enable` (butuh password lagi) →\n   dapat `biometric_token` yang disimpan aplikasi di Android Keystore.\n   Login berikutnya cukup sidik jari lewat `POST /api/auth/biometric/login`.\n\n### Aturan angka\nSemua nilai uang adalah **rupiah bulat** (integer, tanpa desimal).\nPersentase dikirim sebagai angka desimal biasa, mis. `1.75` untuk 1,75%.",
        "contact": {
            "email": "bayuapriansah10@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/kegiatan/{id}/lampiran": {
            "get": {
                "tags": [
                    "Lampiran"
                ],
                "summary": "Daftar lampiran kegiatan",
                "description": "Daftar bukti yang terlampir pada satu kegiatan -- umumnya foto struk belanja bahan baku.",
                "operationId": "lampiranIndex",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lampiran"
                ],
                "summary": "Unggah lampiran / foto struk",
                "description": "Mengunggah bukti belanja. Dikirim sebagai multipart/form-data, sehingga bisa berasal dari pemilih berkas maupun langsung dari kamera perangkat. Berkas yang sama persis (hash identik) ditolak agar tidak terunggah dua kali.",
                "operationId": "lampiranStore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "berkas"
                                ],
                                "properties": {
                                    "berkas": {
                                        "description": "JPG, PNG, WEBP, HEIC, atau PDF. Maksimal 8 MB.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "konteks": {
                                        "type": "string",
                                        "example": "biaya_pelaksanaan",
                                        "enum": [
                                            "biaya_pelaksanaan",
                                            "administrasi",
                                            "lain"
                                        ]
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Terunggah"
                    },
                    "409": {
                        "description": "Berkas identik sudah pernah diunggah"
                    },
                    "422": {
                        "description": "Berkas tidak valid"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lampiran/{id}/berkas": {
            "get": {
                "tags": [
                    "Lampiran"
                ],
                "summary": "Ambil berkas lampiran",
                "description": "Mengunduh isi berkas lampiran. Berkas disimpan di disk privat, jadi hanya bisa diambil lewat endpoint ini yang memeriksa token -- bukan lewat URL publik yang bisa ditebak.",
                "operationId": "lampiranBerkas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Isi berkas",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Berkas tidak ada di penyimpanan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lampiran/{id}": {
            "delete": {
                "tags": [
                    "Lampiran"
                ],
                "summary": "Hapus lampiran",
                "description": "Menghapus lampiran. Berkas fisiknya baru benar-benar dibuang saat penghapusan permanen, sehingga penghapusan tidak sengaja masih bisa dipulihkan.",
                "operationId": "lampiranDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Terhapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan access_token dari endpoint login (tanpa kata \"Bearer\").",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}